+++ vpin-studio-rest-client/src/main/java/de/mephisto/vpin/restclient/highscores/DefaultHighscoresTitles.java
@@ -1,9 +1,0 @@
-package de.mephisto.vpin.restclient.highscores;
-
-import java.util.Arrays;
-import java.util.List;
-
-public interface DefaultHighscoresTitles {
-
-  List<String> DEFAULT_TITLES = Arrays.asList("MASTER MAGICIAN", "CHAMPION", "GRAND CHAMPION", "WORLD RECORD", "GREATEST VAMPIRE HUNTER", "GREATEST TIME LORD", "RIVER MASTER", "CLUB CHAMPION", "HIGHEST SCORES");
-}